refactor/cuda: implement the plugin traits generically against the Cuda backend
This allows the tests/downstream crates to be generic. For Cuda, this doesn't
expand the types allowed, so only `f32` and `f64` are allowed (just as before).
